requirements.txt
protobuf>=3.19.4
googleapis-common-protos>=1.56.0
